Feature/user

See merge request erkhuy/laravel_manage_student!11

+++ app/Http/Controllers/Admins/ClassroomController.php
@@ -1,8 +1,6 @@
 <?php
 
-
 namespace App\Http\Controllers\Admins;
-
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Classrooms\CreateClassroomRequest;
 use App\Http\Requests\Classrooms\UpdateClassroomRequest;
@@ -21,6 +19,7 @@ class ClassroomController extends Controller
 
     public function index(Request $request)
     {
+        $this->authorize('view-classroom');
         $data = $this->classroomRepository->search($request->only(['name', 'subject', 'faculty']));
         return view('backends.classrooms.index')->with([
             'classrooms' => $data['classrooms'],
@@ -32,6 +31,7 @@ class ClassroomController extends Controller
 
     public function create()
     {
+        $this->authorize('create-classroom');
         $data=$this->classroomRepository->create();
         return view('backends.classrooms.create')->with(['faculties'=>$data['faculty'], 'subjects'=>$data['subject']]);
     }
@@ -54,6 +54,7 @@ class ClassroomController extends Controller
 
     public function edit($id)
     {
+        $this->authorize('edit-classroom');
         $data = $this->classroomRepository->edit($id);
         return view('backends.classrooms.edit')->with(['faculties'=>$data['faculty'], 'subjects'=>$data['subject'],'classroom'=>$data['classroom']]);
 
@@ -68,6 +69,7 @@ class ClassroomController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('destroy-classroom');
         return response()->json([
             'status' => 204,
             'message' => $this->classroomRepository->destroy($id),
